+++ src/DDTrace/Configuration/EnvVariableRegistry.php
@@ -13,8 +13,17 @@ class EnvVariableRegistry implements Registry
      */
     private $registry;
 
-    public function __construct()
+    /**
+     * @var string
+     */
+    private $prefix;
+
+    /**
+     * @param string $prefix
+     */
+    public function __construct($prefix = 'DD_')
     {
+        $this->prefix = $prefix;
         $this->registry = [];
     }
 
@@ -24,9 +33,9 @@ class EnvVariableRegistry implements Registry
      * @param string $key
      * @return string|null
      */
-    public static function get($key)
+    protected function get($key)
     {
-        $value = getenv(self::convertKeyToEnvVariableName($key));
+        $value = getenv($this->convertKeyToEnvVariableName($key));
         if (false === $value) {
             return null;
         }
@@ -41,7 +50,7 @@ class EnvVariableRegistry implements Registry
         if (isset($this->registry[$key])) {
             return $this->registry[$key];
         }
-        $value = self::get($key);
+        $value = $this->get($key);
         if (null !== $value) {
             return $this->registry[$key] = $value;
         }
@@ -57,7 +66,7 @@ class EnvVariableRegistry implements Registry
             return $this->registry[$key];
         }
 
-        $value = self::get($key);
+        $value = $this->get($key);
         if (null === $value) {
             return $default;
         }
@@ -80,7 +89,7 @@ class EnvVariableRegistry implements Registry
     public function floatValue($key, $default, $min = null, $max = null)
     {
         if (!isset($this->registry[$key])) {
-            $value = self::get($key);
+            $value = $this->get($key);
             $value = strtolower($value);
             if (is_numeric($value)) {
                 $floatValue = (float)$value;
@@ -116,7 +125,7 @@ class EnvVariableRegistry implements Registry
         }
 
         $default = [];
-        $value = self::get($key);
+        $value = $this->get($key);
 
         if (null === $value) {
             return $default;
@@ -151,7 +160,7 @@ class EnvVariableRegistry implements Registry
     public function inArray($key, $name)
     {
         if (!isset($this->registry[$key])) {
-            $value = self::get($key);
+            $value = $this->get($key);
             if (null !== $value) {
                 $disabledIntegrations = explode(',', $value);
                 $this->registry[$key] = array_map(function ($entry) {
@@ -173,8 +182,8 @@ class EnvVariableRegistry implements Registry
      * @param string $key
      * @return string
      */
-    private static function convertKeyToEnvVariableName($key)
+    private function convertKeyToEnvVariableName($key)
     {
-        return 'DD_' . strtoupper(str_replace('.', '_', trim($key)));
+        return $this->prefix . strtoupper(str_replace('.', '_', trim($key)));
     }
 }