+++ src/DDTrace/Integrations/DefaultIntegrationConfiguration.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace DDTrace\Integrations;
+
+/**
+ * Default integration configuration object for integrations not having any specific param. Put here
+ * config options that are common to ALL integrations.
+ */
+class DefaultIntegrationConfiguration extends AbstractIntegrationConfiguration
+{
+}