+++ src/DDTrace/Integrations/Eloquent/EloquentIntegration.php
@@ -3,14 +3,23 @@
 namespace DDTrace\Integrations\Eloquent;
 
 use DDTrace\Integrations\Integration;
+use DDTrace\Integrations\AbstractIntegration;
 use DDTrace\Tag;
 use DDTrace\Type;
 use DDTrace\Util\TryCatchFinally;
 use DDTrace\GlobalTracer;
 
-class EloquentIntegration
+class EloquentIntegration extends AbstractIntegration
 {
     const NAME = 'eloquent';
+
+    /**
+     * @return string The integration name.
+     */
+    public function getName()
+    {
+        return self::NAME;
+    }
 
     public static function load()
     {