+++ src/DDTrace/Integrations/Memcached/MemcachedIntegration.php
@@ -3,6 +3,7 @@
 namespace DDTrace\Integrations\Memcached;
 
 use DDTrace\Integrations\Integration;
+use DDTrace\Integrations\AbstractIntegration;
 use DDTrace\Obfuscation;
 use DDTrace\Tag;
 use DDTrace\Type;
@@ -23,9 +24,17 @@ use DDTrace\GlobalTracer;
  * might be different for each key. setMultiByKey does, since you're pinning a
  * specific server.
  */
-class MemcachedIntegration
+class MemcachedIntegration extends AbstractIntegration
 {
     const NAME = 'memcached';
+
+    /**
+     * @return string The integration name.
+     */
+    public function getName()
+    {
+        return self::NAME;
+    }
 
     public static function load()
     {