+++ src/DDTrace/Integrations/Mongo/MongoIntegration.php
@@ -2,15 +2,24 @@
 
 namespace DDTrace\Integrations\Mongo;
 
+use DDTrace\Integrations\AbstractIntegration;
 use DDTrace\Span;
 use DDTrace\Tag;
 use DDTrace\Type;
 use DDTrace\Integrations\Integration;
 use DDTrace\Util\Versions;
 
-final class MongoIntegration
+final class MongoIntegration extends AbstractIntegration
 {
     const NAME = 'mongo';
+
+    /**
+     * @return string The integration name.
+     */
+    public function getName()
+    {
+        return self::NAME;
+    }
 
     public static function load()
     {