+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file - [read more
 
 ### Added
 - Disable request_init_hook functionality in presence of blacklisted modules #345
+- Integration-level configuration #354
 
 ### Fixed
 - Symfony template rendering spans #359